Fixed a bug when multiple expanded choices would render unchecked because of the Form Framework's strict type checking.

+++ src/Symfony/Bridge/Doctrine/Form/DataTransformer/EntitiesToArrayTransformer.php
@@ -57,7 +57,7 @@ class EntitiesToArrayTransformer implements DataTransformerInterface
             }
         } else {
             foreach ($collection as $entity) {
-                $array[] = current($this->choiceList->getIdentifierValues($entity));
+                $array[] = (int) current($this->choiceList->getIdentifierValues($entity));
             }
         }
